breaking changes: make *bool usable
+ various fixes for js/wasm
+ fix QT_DEBUG_QML for android(-emulator)
+ make os.Args usable in android(-emulator) and ios(-simulator)

+++ qt.go
@@ -126,7 +126,7 @@ func CountSignals() (c int) {
 	return
 }
 
-func GoBoolToInt(b bool) int {
+func GoBoolToInt(b bool) int8 {
 	if b {
 		return 1
 	}
